Fix int dividing causing wrong numbers to be drawn

+++ src/main/java/modularforcefields/client/screen/ScreenFortronFieldProjector.java
@@ -34,7 +34,7 @@ public class ScreenFortronFieldProjector extends GenericScreen<ContainerFortronF
 		addComponent(new ScreenComponentMultiLabel(0, 0, matrixStack -> {
 			if (menu.getUnsafeHost() instanceof TileFortronFieldProjector projector) {
 				matrixStack.drawString(font, MFFSTextUtils.gui("fortrondevice.linked", projector.getConnections()), 25, 115, 4210752, false);
-				matrixStack.drawString(font, MFFSTextUtils.gui("fortrondevice.usage", ChatFormatter.getChatDisplayShort(projector.getFortronUse() / 1000 * 20, DisplayUnit.BUCKETS)).append(" / s"), 25, 105, 4210752, false);
+				matrixStack.drawString(font, MFFSTextUtils.gui("fortrondevice.usage", ChatFormatter.getChatDisplayShort((int)(projector.getFortronUse() / 1000.0 * 20), DisplayUnit.BUCKETS)).append(" / s"), 25, 105, 4210752, false);
 				matrixStack.drawString(font, MFFSTextUtils.gui("fortrondevice.frequency", projector.getFrequency()), 25, 95, 4210752, false);
 				matrixStack.drawString(font, MFFSTextUtils.gui("fieldprojector.status", projector.getStatus().name()), 8, 130, 4210752, false);
 			}
